OWB-1061 mark ApplicationContext as already destroyed
We cannot immediately kill out the whole ApplicationContext and not even set it to inactive
as it is needed for delivering @BeforeShutdown events.


git-svn-id: https://svn.apache.org/repos/asf/openwebbeans/trunk@1681661 13f79535-47bb-0310-9956-ffa450edef68

+++ webbeans-impl/src/main/java/org/apache/webbeans/corespi/se/DefaultContextsService.java
@@ -296,7 +296,7 @@ public class DefaultContextsService extends AbstractContextsService
     
     private void startApplicationContext()
     {
-        if (applicationContext != null)
+        if (applicationContext != null && !applicationContext.isDestroyed())
         {
             // applicationContext is already started
             return;
@@ -368,7 +368,7 @@ public class DefaultContextsService extends AbstractContextsService
     
     private void stopApplicationContext()
     {
-        if(applicationContext != null)
+        if(applicationContext != null && !applicationContext.isDestroyed())
         {
             applicationContext.destroy();
